requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
langgraph>=0.0.11
tavily-python>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0